fix: add css style for row when context menu display
